Refactor handlers to extract reusable functions
