+++ subgraphs/governance/openzeppelin-governor-forks/protocols/truefi-governance/src/dao-governor.ts
@@ -1,4 +1,4 @@
-import { log } from "@graphprotocol/graph-ts";
+import { Address } from "@graphprotocol/graph-ts";
 import {
   ProposalCanceled,
   ProposalCreated,
@@ -12,107 +12,46 @@ import {
   VotingDelaySet,
   VotingPeriodSet,
 } from "../../../generated/DaoGovernor/DaoGovernor";
-import { Vote } from "../../../generated/schema";
 import {
-  getOrCreateProposal,
-  ProposalState,
-  getGovernance,
-  BIGINT_ONE,
-  addressesToStrings,
-  getOrCreateDelegate,
-  getGovernanceFramework,
-  getVoteChoiceByValue,
-  BIGINT_ZERO,
-} from "./helpers";
+  _handleProposalCreated,
+  _handleProposalCanceled,
+  _handleProposalExecuted,
+  _handleProposalQueued,
+  _handleVoteCast,
+} from "../../../src/handlers";
+import { DaoGovernor } from "../../../generated/DaoGovernor/DaoGovernor";
+import { GovernanceFramework } from "../../../generated/schema";
 
 export function handleProposalCanceled(event: ProposalCanceled): void {
-  let proposal = getOrCreateProposal(event.params.proposalId.toString());
-  proposal.state = ProposalState.CANCELED;
-  proposal.cancellationBlock = event.block.number;
-  proposal.cancellationTime = event.block.timestamp;
-  proposal.save();
-
-  // Update governance proposal state counts
-  const governance = getGovernance();
-  governance.proposalsCanceled = governance.proposalsCanceled.plus(BIGINT_ONE);
-  governance.save();
+  _handleProposalCanceled({ proposalId: event.params.proposalId }, event);
 }
 
 export function handleProposalCreated(event: ProposalCreated): void {
-  let proposal = getOrCreateProposal(event.params.proposalId.toString());
-  let proposer = getOrCreateDelegate(
-    event.params.proposer.toHexString(),
-    false
+  _handleProposalCreated(
+    {
+      proposalId: event.params.proposalId,
+      proposer: event.params.proposer,
+      targets: event.params.targets,
+      values: event.params.values,
+      signatures: event.params.signatures,
+      calldatas: event.params.calldatas,
+      startBlock: event.params.startBlock,
+      endBlock: event.params.endBlock,
+      description: event.params.description,
+    },
+    event
   );
-
-  // Checking if the proposer was a delegate already accounted for, if not we should log an error
-  // since it shouldn't be possible for a delegate to propose anything without first being "created"
-  if (proposer == null) {
-    log.error(
-      "Delegate participant {} not found on ProposalCreated. tx_hash: {}",
-      [
-        event.params.proposer.toHexString(),
-        event.transaction.hash.toHexString(),
-      ]
-    );
-  }
-
-  // Creating it anyway since we will want to account for this event data, even though it should've never happened
-  proposer = getOrCreateDelegate(event.params.proposer.toHexString());
-
-  proposal.proposer = proposer.id;
-  proposal.againstVotes = BIGINT_ZERO;
-  proposal.forVotes = BIGINT_ZERO;
-  proposal.abstainVotes = BIGINT_ZERO;
-  proposal.targets = addressesToStrings(event.params.targets);
-  proposal.values = event.params.values;
-  proposal.signatures = event.params.signatures;
-  proposal.calldatas = event.params.calldatas;
-  proposal.creationBlock = event.block.number;
-  proposal.creationTime = event.block.timestamp;
-  proposal.startBlock = event.params.startBlock;
-  proposal.endBlock = event.params.endBlock;
-  proposal.description = event.params.description;
-  proposal.state =
-    event.block.number >= proposal.startBlock
-      ? ProposalState.ACTIVE
-      : ProposalState.PENDING;
-  proposal.governanceFramework = event.address.toHexString();
-  proposal.save();
-
-  // Increment gov proposal count
-  const governance = getGovernance();
-  governance.proposals = governance.proposals.plus(BIGINT_ONE);
-  governance.save();
 }
 
 export function handleProposalExecuted(event: ProposalExecuted): void {
-  // Update proposal status + execution metadata
-  let proposal = getOrCreateProposal(event.params.proposalId.toString());
-  proposal.state = ProposalState.EXECUTED;
-  proposal.executionETA = null;
-  proposal.executionBlock = event.block.number;
-  proposal.executionTime = event.block.timestamp;
-  proposal.save();
-
-  // Update governance proposal state counts
-  let governance = getGovernance();
-  governance.proposalsQueued = governance.proposalsQueued.minus(BIGINT_ONE);
-  governance.proposalsExecuted = governance.proposalsExecuted.plus(BIGINT_ONE);
-  governance.save();
+  _handleProposalExecuted({ proposalId: event.params.proposalId }, event);
 }
 
 export function handleProposalQueued(event: ProposalQueued): void {
-  // Update proposal status + execution metadata
-  let proposal = getOrCreateProposal(event.params.proposalId.toString());
-  proposal.state = ProposalState.QUEUED;
-  proposal.executionETA = event.params.eta;
-  proposal.save();
-
-  // Update governance proposal state counts
-  let governance = getGovernance();
-  governance.proposalsQueued = governance.proposalsQueued.plus(BIGINT_ONE);
-  governance.save();
+  _handleProposalQueued(
+    { proposalId: event.params.proposalId, eta: event.params.eta },
+    event
+  );
 }
 
 export function handleProposalThresholdSet(event: ProposalThresholdSet): void {
@@ -136,75 +75,29 @@ export function handleTimelockChange(event: TimelockChange): void {
 }
 
 export function handleVoteCast(event: VoteCast): void {
-  const proposalId = event.params.proposalId.toString();
-  const voterAddress = event.params.voter.toHexString();
-
-  let voteId = voterAddress.concat("-").concat(proposalId);
-  let vote = new Vote(voteId);
-  vote.proposal = proposalId;
-  vote.voter = voterAddress;
-  vote.weight = event.params.weight;
-  vote.reason = event.params.reason;
-
-  // Retrieve enum string key by value (0 = Against, 1 = For, 2 = Abstain)
-  vote.choice = getVoteChoiceByValue(event.params.support);
-  vote.save();
-
-  let proposal = getOrCreateProposal(proposalId);
-  if (proposal.state == ProposalState.PENDING) {
-    proposal.state = ProposalState.ACTIVE;
-  }
-
-  // Increment respective vote choice counts
-  if (event.params.support === 0) {
-    proposal.againstVotes = proposal.againstVotes.plus(BIGINT_ONE);
-  } else if (event.params.support === 1) {
-    proposal.forVotes = proposal.forVotes.plus(BIGINT_ONE);
-  } else if (event.params.support === 2) {
-    proposal.abstainVotes = proposal.abstainVotes.plus(BIGINT_ONE);
-  }
-  proposal.save();
-
-  // Add 1 to participant's proposal voting count
-  let voter = getOrCreateDelegate(voterAddress);
-  voter.numberVotes = voter.numberVotes + 1;
-  voter.save();
+  _handleVoteCast(
+    {
+      proposalId: event.params.proposalId,
+      voter: event.params.voter,
+      weight: event.params.weight,
+      reason: event.params.reason,
+      support: event.params.support,
+    },
+    event
+  );
 }
 // Treat VoteCastWithParams same as VoteCast
 export function handleVoteCastWithParams(event: VoteCastWithParams): void {
-  const proposalId = event.params.proposalId.toString();
-  const voterAddress = event.params.voter.toHexString();
-
-  let voteId = voterAddress.concat("-").concat(proposalId);
-  let vote = new Vote(voteId);
-  vote.proposal = proposalId;
-  vote.voter = voterAddress;
-  vote.weight = event.params.weight;
-  vote.reason = event.params.reason;
-
-  // Retrieve enum string key by value (0 = Against, 1 = For, 2 = Abstain)
-  vote.choice = getVoteChoiceByValue(event.params.support);
-  vote.save();
-
-  let proposal = getOrCreateProposal(proposalId);
-  if (proposal.state == ProposalState.PENDING) {
-    proposal.state = ProposalState.ACTIVE;
-  }
-
-  // Increment respective vote choice counts
-  if (event.params.support === 0) {
-    proposal.againstVotes = proposal.againstVotes.plus(BIGINT_ONE);
-  } else if (event.params.support === 1) {
-    proposal.forVotes = proposal.forVotes.plus(BIGINT_ONE);
-  } else if (event.params.support === 2) {
-    proposal.abstainVotes = proposal.abstainVotes.plus(BIGINT_ONE);
-  }
-  proposal.save();
-
-  // Add 1 to participant's proposal voting count
-  let voter = getOrCreateDelegate(voterAddress);
-  voter.numberVotes = voter.numberVotes + 1;
-  voter.save();
+  _handleVoteCast(
+    {
+      proposalId: event.params.proposalId,
+      voter: event.params.voter,
+      weight: event.params.weight,
+      reason: event.params.reason,
+      support: event.params.support,
+    },
+    event
+  );
 }
 
 export function handleVotingDelaySet(event: VotingDelaySet): void {
@@ -217,4 +110,30 @@ export function handleVotingPeriodSet(event: VotingPeriodSet): void {
   let governanceFramework = getGovernanceFramework(event.address.toHexString());
   governanceFramework.votingPeriod = event.params.newVotingPeriod;
   governanceFramework.save();
+}
+
+// Helper function that imports and binds the contract
+function getGovernanceFramework(contractAddress: string): GovernanceFramework {
+  let governanceFramework = GovernanceFramework.load(contractAddress);
+
+  if (!governanceFramework) {
+    governanceFramework = new GovernanceFramework(contractAddress);
+    let contract = DaoGovernor.bind(Address.fromString(contractAddress));
+
+    governanceFramework.name = contract.name();
+    governanceFramework.type = "OZGovernor";
+    governanceFramework.version = contract.version();
+
+    governanceFramework.contractAddress = contractAddress;
+    governanceFramework.tokenAddress = contract.timelock().toHexString();
+    governanceFramework.timelockAddress = contract.token().toHexString();
+
+    governanceFramework.votingDelay = contract.votingDelay();
+    governanceFramework.votingPeriod = contract.votingPeriod();
+    governanceFramework.proposalThreshold = contract.proposalThreshold();
+    governanceFramework.quorumNumerator = contract.quorumNumerator();
+    governanceFramework.quorumDenominator = contract.quorumDenominator();
+  }
+
+  return governanceFramework;
 }
