--- subgraphs/governance/openzeppelin-governor-forks/protocols/truefi-governance/src/helpers.ts
@@ -1,195 +1,0 @@
-import {
-  Address,
-  BigDecimal,
-  BigInt,
-  Bytes,
-  log,
-} from "@graphprotocol/graph-ts";
-import { DaoGovernor } from "../../../generated/DaoGovernor/DaoGovernor";
-import {
-  Delegate,
-  Governance,
-  GovernanceFramework,
-  Proposal,
-  TokenHolder,
-} from "../../../generated/schema";
-
-export const ZERO_ADDRESS = "0x0000000000000000000000000000000000000000";
-export const BIGINT_ZERO = BigInt.fromI32(0);
-export const BIGINT_ONE = BigInt.fromI32(1);
-export const BIGINT_FIVE = BigInt.fromI32(5);
-export const BIGDECIMAL_ZERO = new BigDecimal(BIGINT_ZERO);
-
-export const GOVERNANCE_NAME = "OZGovernor";
-export namespace ProposalState {
-  export const PENDING = "PENDING";
-  export const ACTIVE = "ACTIVE";
-  export const CANCELED = "CANCELED";
-  export const DEFEATED = "DEFEATED";
-  export const SUCCEEDED = "SUCCEEDED";
-  export const QUEUED = "QUEUED";
-  export const EXPIRED = "EXPIRED";
-  export const EXECUTED = "EXECUTED";
-}
-
-export namespace VoteChoice {
-  export const AGAINST_VALUE = 0;
-  export const FOR_VALUE = 1;
-  export const ABSTAIN_VALUE = 2;
-  export const AGAINST = "AGAINST";
-  export const FOR = "FOR";
-  export const ABSTAIN = "ABSTAIN";
-}
-
-export function getVoteChoiceByValue(choiceValue: number): string {
-  if (choiceValue === VoteChoice.ABSTAIN_VALUE) {
-    return VoteChoice.AGAINST;
-  } else if (choiceValue === VoteChoice.FOR_VALUE) {
-    return VoteChoice.FOR;
-  } else if (choiceValue === VoteChoice.ABSTAIN_VALUE) {
-    return VoteChoice.ABSTAIN;
-  } else {
-    // Case that shouldn't happen
-    log.error("Voting choice of {} does not exist", [choiceValue.toString()]);
-    return VoteChoice.ABSTAIN;
-  }
-}
-
-export function toDecimal(value: BigInt, decimals: number = 18): BigDecimal {
-  return value.divDecimal(
-    BigInt.fromI32(10)
-      .pow(<u8>decimals)
-      .toBigDecimal()
-  );
-}
-
-export function addressesToStrings(addresses: Address[]): Array<string> {
-  const byteAddresses = new Array<string>();
-  for (let i = 0; i < addresses.length; i++) {
-    byteAddresses.push(addresses[i].toHexString());
-  }
-  return byteAddresses;
-}
-
-export function getGovernance(): Governance {
-  let governance = Governance.load(GOVERNANCE_NAME);
-
-  if (!governance) {
-    governance = new Governance(GOVERNANCE_NAME);
-    governance.proposals = BIGINT_ZERO;
-    governance.currentTokenHolders = BIGINT_ZERO;
-    governance.totalTokenHolders = BIGINT_ZERO;
-    governance.currentDelegates = BIGINT_ZERO;
-    governance.totalDelegates = BIGINT_ZERO;
-    governance.delegatedVotesRaw = BIGINT_ZERO;
-    governance.delegatedVotes = BIGDECIMAL_ZERO;
-    governance.proposalsQueued = BIGINT_ZERO;
-    governance.proposalsExecuted = BIGINT_ZERO;
-    governance.proposalsCanceled = BIGINT_ZERO;
-  }
-
-  return governance;
-}
-
-export function getGovernanceFramework(
-  contractAddress: string
-): GovernanceFramework {
-  let governanceFramework = GovernanceFramework.load(contractAddress);
-
-  if (!governanceFramework) {
-    governanceFramework = new GovernanceFramework(contractAddress);
-    let contract = DaoGovernor.bind(Address.fromString(contractAddress));
-
-    governanceFramework.name = contract.name();
-    governanceFramework.type = "OZGovernor";
-    governanceFramework.version = contract.version();
-
-    governanceFramework.contractAddress = contractAddress;
-    governanceFramework.tokenAddress = contract.timelock().toHexString();
-    governanceFramework.timelockAddress = contract.token().toHexString();
-
-    governanceFramework.votingDelay = contract.votingDelay();
-    governanceFramework.votingPeriod = contract.votingPeriod();
-    governanceFramework.proposalThreshold = contract.proposalThreshold();
-    governanceFramework.quorumNumerator = contract.quorumNumerator();
-    governanceFramework.quorumDenominator = contract.quorumDenominator();
-  }
-
-  return governanceFramework;
-}
-
-export function getOrCreateProposal(
-  id: string,
-  createIfNotFound: boolean = true,
-  save: boolean = false
-): Proposal {
-  let proposal = Proposal.load(id);
-
-  if (!proposal && createIfNotFound) {
-    proposal = new Proposal(id);
-    if (save) {
-      proposal.save();
-    }
-
-    let governance = getGovernance();
-    governance.proposals = governance.proposals.plus(BIGINT_ONE);
-    governance.save();
-  }
-
-  return proposal as Proposal;
-}
-
-export function getOrCreateDelegate(
-  address: string,
-  createIfNotFound: boolean = true,
-  save: boolean = true
-): Delegate {
-  let delegate = Delegate.load(address);
-
-  if (!delegate && createIfNotFound) {
-    delegate = new Delegate(address);
-    delegate.delegatedVotesRaw = BIGINT_ZERO;
-    delegate.delegatedVotes = BIGDECIMAL_ZERO;
-    delegate.tokenHoldersRepresentedAmount = 0;
-    delegate.numberVotes = 0;
-    if (save) {
-      delegate.save();
-    }
-
-    if (address != ZERO_ADDRESS) {
-      let governance = getGovernance();
-      governance.totalDelegates = governance.totalDelegates.plus(BIGINT_ONE);
-      governance.save();
-    }
-  }
-
-  return delegate as Delegate;
-}
-
-export function getOrCreateTokenHolder(
-  address: string,
-  createIfNotFound: boolean = true,
-  save: boolean = true
-): TokenHolder {
-  let tokenHolder = TokenHolder.load(address);
-
-  if (!tokenHolder && createIfNotFound) {
-    tokenHolder = new TokenHolder(address);
-    tokenHolder.tokenBalanceRaw = BIGINT_ZERO;
-    tokenHolder.tokenBalance = BIGDECIMAL_ZERO;
-    tokenHolder.totalTokensHeldRaw = BIGINT_ZERO;
-    tokenHolder.totalTokensHeld = BIGDECIMAL_ZERO;
-    if (save) {
-      tokenHolder.save();
-    }
-
-    if (address != ZERO_ADDRESS) {
-      let governance = getGovernance();
-      governance.totalTokenHolders =
-        governance.totalTokenHolders.plus(BIGINT_ONE);
-      governance.save();
-    }
-  }
-
-  return tokenHolder as TokenHolder;
-}
