+++ subgraphs/governance/openzeppelin-governor-forks/src/constants.ts
@@ -1,33 +1,28 @@
 import { BigDecimal, BigInt } from "@graphprotocol/graph-ts";
 
-////////////////////////
-///// Schema Enums /////
-////////////////////////
+export const ZERO_ADDRESS = "0x0000000000000000000000000000000000000000";
+export const BIGINT_ZERO = BigInt.fromI32(0);
+export const BIGINT_ONE = BigInt.fromI32(1);
+export const BIGINT_FIVE = BigInt.fromI32(5);
+export const BIGDECIMAL_ZERO = new BigDecimal(BIGINT_ZERO);
 
-// The network names corresponding to the Network enum in the schema.
-// They also correspond to the ones in `dataSource.network()` after converting to lower case.
-// See below for a complete list:
-// https://thegraph.com/docs/en/hosted-service/what-is-hosted-service/#supported-networks-on-the-hosted-service
-export namespace Network {
-  export const ARBITRUM_ONE = "ARBITRUM_ONE";
-  export const ARWEAVE_MAINNET = "ARWEAVE_MAINNET";
-  export const AURORA = "AURORA";
-  export const AVALANCHE = "AVALANCHE";
-  export const BOBA = "BOBA";
-  export const BSC = "BSC"; // aka BNB Chain
-  export const CELO = "CELO";
-  export const COSMOS = "COSMOS";
-  export const CRONOS = "CRONOS";
-  export const MAINNET = "MAINNET"; // Ethereum mainnet
-  export const FANTOM = "FANTOM";
-  export const FUSE = "FUSE";
-  export const HARMONY = "HARMONY";
-  export const JUNO = "JUNO";
-  export const MOONBEAM = "MOONBEAM";
-  export const MOONRIVER = "MOONRIVER";
-  export const NEAR_MAINNET = "NEAR_MAINNET";
-  export const OPTIMISM = "OPTIMISM";
-  export const OSMOSIS = "OSMOSIS";
-  export const MATIC = "MATIC"; // aka Polygon
-  export const XDAI = "XDAI"; // aka Gnosis Chain
+export const GOVERNANCE_NAME = "OZGovernor";
+export namespace ProposalState {
+  export const PENDING = "PENDING";
+  export const ACTIVE = "ACTIVE";
+  export const CANCELED = "CANCELED";
+  export const DEFEATED = "DEFEATED";
+  export const SUCCEEDED = "SUCCEEDED";
+  export const QUEUED = "QUEUED";
+  export const EXPIRED = "EXPIRED";
+  export const EXECUTED = "EXECUTED";
+}
+
+export namespace VoteChoice {
+  export const AGAINST_VALUE = 0;
+  export const FOR_VALUE = 1;
+  export const ABSTAIN_VALUE = 2;
+  export const AGAINST = "AGAINST";
+  export const FOR = "FOR";
+  export const ABSTAIN = "ABSTAIN";
 }
