+++ subgraphs/governance/openzeppelin-governor-forks/src/handlers.ts
@@ -1,0 +1,301 @@
+import {
+  Address,
+  BigDecimal,
+  BigInt,
+  Bytes,
+  ethereum,
+  log,
+} from "@graphprotocol/graph-ts";
+import {
+  BIGDECIMAL_ZERO,
+  BIGINT_ONE,
+  BIGINT_ZERO,
+  GOVERNANCE_NAME,
+  ProposalState,
+  VoteChoice,
+  ZERO_ADDRESS,
+} from "./constants";
+import {
+  Delegate,
+  Governance,
+  Proposal,
+  TokenHolder,
+  Vote,
+} from "../generated/schema";
+
+export function toDecimal(value: BigInt, decimals: number = 18): BigDecimal {
+  return value.divDecimal(
+    BigInt.fromI32(10)
+      .pow(<u8>decimals)
+      .toBigDecimal()
+  );
+}
+export function addressesToStrings(addresses: Address[]): Array<string> {
+  const byteAddresses = new Array<string>();
+  for (let i = 0; i < addresses.length; i++) {
+    byteAddresses.push(addresses[i].toHexString());
+  }
+  return byteAddresses;
+}
+
+export function getVoteChoiceByValue(choiceValue: number): string {
+  if (choiceValue === VoteChoice.ABSTAIN_VALUE) {
+    return VoteChoice.AGAINST;
+  } else if (choiceValue === VoteChoice.FOR_VALUE) {
+    return VoteChoice.FOR;
+  } else if (choiceValue === VoteChoice.ABSTAIN_VALUE) {
+    return VoteChoice.ABSTAIN;
+  } else {
+    // Case that shouldn't happen
+    log.error("Voting choice of {} does not exist", [choiceValue.toString()]);
+    return VoteChoice.ABSTAIN;
+  }
+}
+
+export function getGovernance(): Governance {
+  let governance = Governance.load(GOVERNANCE_NAME);
+
+  if (!governance) {
+    governance = new Governance(GOVERNANCE_NAME);
+    governance.proposals = BIGINT_ZERO;
+    governance.currentTokenHolders = BIGINT_ZERO;
+    governance.totalTokenHolders = BIGINT_ZERO;
+    governance.currentDelegates = BIGINT_ZERO;
+    governance.totalDelegates = BIGINT_ZERO;
+    governance.delegatedVotesRaw = BIGINT_ZERO;
+    governance.delegatedVotes = BIGDECIMAL_ZERO;
+    governance.proposalsQueued = BIGINT_ZERO;
+    governance.proposalsExecuted = BIGINT_ZERO;
+    governance.proposalsCanceled = BIGINT_ZERO;
+  }
+
+  return governance;
+}
+
+export function getOrCreateProposal(
+  id: string,
+  createIfNotFound: boolean = true,
+  save: boolean = false
+): Proposal {
+  let proposal = Proposal.load(id);
+
+  if (!proposal && createIfNotFound) {
+    proposal = new Proposal(id);
+    if (save) {
+      proposal.save();
+    }
+
+    let governance = getGovernance();
+    governance.proposals = governance.proposals.plus(BIGINT_ONE);
+    governance.save();
+  }
+
+  return proposal as Proposal;
+}
+
+export function getOrCreateDelegate(
+  address: string,
+  createIfNotFound: boolean = true,
+  save: boolean = true
+): Delegate {
+  let delegate = Delegate.load(address);
+
+  if (!delegate && createIfNotFound) {
+    delegate = new Delegate(address);
+    delegate.delegatedVotesRaw = BIGINT_ZERO;
+    delegate.delegatedVotes = BIGDECIMAL_ZERO;
+    delegate.tokenHoldersRepresentedAmount = 0;
+    delegate.numberVotes = 0;
+    if (save) {
+      delegate.save();
+    }
+
+    if (address != ZERO_ADDRESS) {
+      let governance = getGovernance();
+      governance.totalDelegates = governance.totalDelegates.plus(BIGINT_ONE);
+      governance.save();
+    }
+  }
+
+  return delegate as Delegate;
+}
+
+export function getOrCreateTokenHolder(
+  address: string,
+  createIfNotFound: boolean = true,
+  save: boolean = true
+): TokenHolder {
+  let tokenHolder = TokenHolder.load(address);
+
+  if (!tokenHolder && createIfNotFound) {
+    tokenHolder = new TokenHolder(address);
+    tokenHolder.tokenBalanceRaw = BIGINT_ZERO;
+    tokenHolder.tokenBalance = BIGDECIMAL_ZERO;
+    tokenHolder.totalTokensHeldRaw = BIGINT_ZERO;
+    tokenHolder.totalTokensHeld = BIGDECIMAL_ZERO;
+    if (save) {
+      tokenHolder.save();
+    }
+
+    if (address != ZERO_ADDRESS) {
+      let governance = getGovernance();
+      governance.totalTokenHolders =
+        governance.totalTokenHolders.plus(BIGINT_ONE);
+      governance.save();
+    }
+  }
+
+  return tokenHolder as TokenHolder;
+}
+
+export interface ProposalCreatedEventParams {
+  proposalId: BigInt;
+  proposer: Address;
+  targets: Address[];
+  values: BigInt[];
+  signatures: string[];
+  calldatas: Bytes[];
+  startBlock: BigInt;
+  endBlock: BigInt;
+  description: string;
+}
+export function _handleProposalCreated(
+  params: ProposalCreatedEventParams,
+  event: ethereum.Event
+): void {
+  let proposal = getOrCreateProposal(params.proposalId.toString());
+  let proposer = getOrCreateDelegate(params.proposer.toHexString(), false);
+
+  // Checking if the proposer was a delegate already accounted for, if not we should log an error
+  // since it shouldn't be possible for a delegate to propose anything without first being "created"
+  if (proposer == null) {
+    log.error(
+      "Delegate participant {} not found on ProposalCreated. tx_hash: {}",
+      [params.proposer.toHexString(), event.transaction.hash.toHexString()]
+    );
+  }
+
+  // Creating it anyway since we will want to account for this event data, even though it should've never happened
+  proposer = getOrCreateDelegate(params.proposer.toHexString());
+
+  proposal.proposer = proposer.id;
+  proposal.againstVotes = BIGINT_ZERO;
+  proposal.forVotes = BIGINT_ZERO;
+  proposal.abstainVotes = BIGINT_ZERO;
+  proposal.targets = addressesToStrings(params.targets);
+  proposal.values = params.values;
+  proposal.signatures = params.signatures;
+  proposal.calldatas = params.calldatas;
+  proposal.creationBlock = event.block.number;
+  proposal.creationTime = event.block.timestamp;
+  proposal.startBlock = params.startBlock;
+  proposal.endBlock = params.endBlock;
+  proposal.description = params.description;
+  proposal.state =
+    event.block.number >= proposal.startBlock
+      ? ProposalState.ACTIVE
+      : ProposalState.PENDING;
+  proposal.governanceFramework = event.address.toHexString();
+  proposal.save();
+
+  // Increment gov proposal count
+  const governance = getGovernance();
+  governance.proposals = governance.proposals.plus(BIGINT_ONE);
+  governance.save();
+}
+
+export function _handleProposalCanceled(
+  params: { proposalId: BigInt },
+  event: ethereum.Event
+): void {
+  let proposal = getOrCreateProposal(params.proposalId.toString());
+  proposal.state = ProposalState.CANCELED;
+  proposal.cancellationBlock = event.block.number;
+  proposal.cancellationTime = event.block.timestamp;
+  proposal.save();
+
+  // Update governance proposal state counts
+  const governance = getGovernance();
+  governance.proposalsCanceled = governance.proposalsCanceled.plus(BIGINT_ONE);
+  governance.save();
+}
+
+export function _handleProposalExecuted(
+  params: { proposalId: BigInt },
+  event: ethereum.Event
+): void {
+  // Update proposal status + execution metadata
+  let proposal = getOrCreateProposal(params.proposalId.toString());
+  proposal.state = ProposalState.EXECUTED;
+  proposal.executionBlock = event.block.number;
+  proposal.executionTime = event.block.timestamp;
+  proposal.save();
+
+  // Update governance proposal state counts
+  let governance = getGovernance();
+  governance.proposalsQueued = governance.proposalsQueued.minus(BIGINT_ONE);
+  governance.proposalsExecuted = governance.proposalsExecuted.plus(BIGINT_ONE);
+  governance.save();
+}
+
+export function _handleProposalQueued(
+  params: { proposalId: BigInt; eta: BigInt },
+  event: ethereum.Event
+): void {
+  // Update proposal status + execution metadata
+  let proposal = getOrCreateProposal(params.proposalId.toString());
+  proposal.state = ProposalState.QUEUED;
+  proposal.executionETA = params.eta;
+  proposal.save();
+
+  // Update governance proposal state counts
+  let governance = getGovernance();
+  governance.proposalsQueued = governance.proposalsQueued.plus(BIGINT_ONE);
+  governance.save();
+}
+
+export interface VoteCastEventParams {
+  proposalId: BigInt;
+  voter: Address;
+  weight: BigInt;
+  reason: string;
+  support: i32;
+}
+export function _handleVoteCast(
+  params: VoteCastEventParams,
+  event: ethereum.Event
+): void {
+  const proposalId = params.proposalId.toString();
+  const voterAddress = params.voter.toHexString();
+
+  let voteId = voterAddress.concat("-").concat(proposalId);
+  let vote = new Vote(voteId);
+  vote.proposal = proposalId;
+  vote.voter = voterAddress;
+  vote.weight = params.weight;
+  vote.reason = params.reason;
+
+  // Retrieve enum string key by value (0 = Against, 1 = For, 2 = Abstain)
+  vote.choice = getVoteChoiceByValue(params.support);
+  vote.save();
+
+  let proposal = getOrCreateProposal(proposalId);
+  if (proposal.state == ProposalState.PENDING) {
+    proposal.state = ProposalState.ACTIVE;
+  }
+
+  // Increment respective vote choice counts
+  if (params.support === 0) {
+    proposal.againstVotes = proposal.againstVotes.plus(BIGINT_ONE);
+  } else if (params.support === 1) {
+    proposal.forVotes = proposal.forVotes.plus(BIGINT_ONE);
+  } else if (params.support === 2) {
+    proposal.abstainVotes = proposal.abstainVotes.plus(BIGINT_ONE);
+  }
+  proposal.save();
+
+  // Add 1 to participant's proposal voting count
+  let voter = getOrCreateDelegate(voterAddress);
+  voter.numberVotes = voter.numberVotes + 1;
+  voter.save();
+}
